moved index.html into root; fixed build config to only package and publish dist
