+++ index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <title>Rolodex</title>
+        <script src="./index.ts" type="module"></script>
+        <style>
+            * {
+                box-sizing: border-box;
+                font-family: sans-serif;
+            }
+
+            body {
+                align-items: center;
+                display: flex;
+                height: 100vh;
+                justify-content: center;
+                width: 100vw;
+            }
+
+            h1 {
+                text-align: left;
+            }
+        </style>
+    </head>
+    <body>
+        <div id="app"></div>
+        <script type="module" src="/src/index.ts"></script>
+    </body>
+</html>
