--- src/index.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>Rolodex</title>
-        <script src="./index.ts" type="module"></script>
-        <style>
-            * {
-                box-sizing: border-box;
-                font-family: sans-serif;
-            }
-
-            body {
-                align-items: center;
-                display: flex;
-                height: 100vh;
-                justify-content: center;
-                width: 100vw;
-            }
-
-            h1 {
-                text-align: left;
-            }
-        </style>
-    </head>
-    <body>
-        <h1>Can you imagine the <rolodex-animation options='["impact", "damage", "surprise"]' transition-timing-function="step-start"></rolodex-animation>?</h1>
-        <!-- <h1> -->
-        <!--     Attention -->
-        <!--     <rolodex-animation interval="1000" transition-duration="0.5" options='["Mac", "Linux"]' transition-timing-function="ease-in"></rolodex-animation> -->
-        <!--     users! -->
-        <!-- </h1> -->
-        <!-- <h1>Github... the most<rolodex-animation options='["amazing", "wonderful", "incredible"]' transition-timing-function="ease-in"></rolodex-animation>GIT repository in the world</h1> -->
-        <!-- <h1>Can you imagine the <rolodex-animation options='["impact", "damage", "surprise"]'></rolodex-animation>?</h1> -->
-        <!-- <h2>Can you imagine the <rolodex-animation options='["impact", "damage", "surprise"]'></rolodex-animation>?</h2> -->
-        <!-- <h3>Can you imagine the <rolodex-animation options='["impact", "damage", "surprise"]'></rolodex-animation>?</h3> -->
-        <!-- <h4>Can you imagine the <rolodex-animation options='["impact", "damage", "surprise"]'></rolodex-animation>?</h4> -->
-        <!-- <h5>Can you imagine the <rolodex-animation options='["impact", "damage", "surprise"]'></rolodex-animation>?</h5> -->
-        <!-- <h6>Can you imagine the <rolodex-animation options='["impact", "damage", "surprise"]'></rolodex-animation>?</h6> -->
-    </body>
-</html>
